Quantile distribution function.

Q Quantile function for multiple distributions.

Usage

Q(
  p,
  dist = "normal",
  lower.tail = TRUE,
  two.sided = FALSE,
  rounding = 2,
  gui = "plot",
  mfrow = c(1, 2),
  type = "both",
  ...
)

Arguments

p

probability. The p argument need have length 1 and value lower then 1.

dist

distribution to use. The default is 'normal'. Options: 'normal', 't-student', 'gumbel', 'binomial', 'poisson', and ....

lower.tail

logical; if TRUE (default), the quantile function is computed; otherwise, the complement of the quantile function (survival function) will be computed. The lower.tail argument will only be valid for two-sided = FALSE.

two.sided

logical. if TRUE (default), the calculation of the quantile function and survival will be presented; otherwise the Q() function will be based according to the lower.tail argument.

rounding

numerical; it represents the number of decimals for calculating the probability.

gui

default is 'plot'; it graphically displays the result of the probability. Others options are: "plot" and "rstudio" and "tcltk".

mfrow

numerical vector. Considering the arguments two.sided = TRUE and type = "both", the default will be to present two graphs (based on CDF and PDF) horizontally for the quantile function, that is, mfrow = c(1, 2) (default).

type

character argument. The default is "both"; the output will display two plots (based on CDF and PDF) to present the result of Q(). The other options are: "cdf" and "pdf".

...

additional parameters according to the chosen distribution.

Value

Q returns the quantile and its graphical representation for a given distribution. The output is a vector.

Details

The expression of quantile function is given by: $$ Q(p)=\inf {x\in \mathbb{R}: p \le F(x)}, $$ where p is the first argument of Q() and x its return value;

Examples

# Attaching package
library(leem)
if (FALSE) { # \dontrun{
Q(p = 0.8, dist = "normal", mean = 200, sd=30)
} # }