Frequency distribution table

Generic function that allows you to tabulate continuous and categorical data (quantitative or qualitative) in frequency distribution. Depending on the nature of the data, they can be grouped into class ranges or not.

Usage

tabfreq(data, ...)

Arguments

data

R object (data structure vector) of class leem. Use new_leem() function.

...

further arguments passed to or from other methods.

Value

The result of tabfreq() is a list. This list has two elements: table and statistics. The first is the data frequency table, and the second represents some useful statistics for methods of leem class.

Examples

# Example 1
library(leem)
x <- rbinom(36, 10, 0.6)
x <- new_leem(x, variable = "discrete")
tabfreq(x)
#> 
#> Table of frequency 
#> Type of variable: discrete
#> 
#>   Groups Fi   Fr Fac1 Fac2 Fp  Fac1p  Fac2p
#> 1      2  4 0.11    4   36 11  11.11 100.00
#> 2      3  2 0.06    6   32  6  16.67  88.89
#> 3      4  6 0.17   12   30 17  33.33  83.33
#> 4      5  7 0.19   19   24 19  52.78  66.67
#> 5      6  6 0.17   25   17 17  69.44  47.22
#> 6      7  3 0.08   28   11  8  77.78  30.56
#> 7      8  5 0.14   33    8 14  91.67  22.22
#> 8      9  3 0.08   36    3  8 100.00   8.33
#> ============================================== 
#> Groups: Discretized grouping 
#> Fi: Absolute frequency 
#> Fr: Relative frequency 
#> Fac1: Cumulative frequency (below) 
#> Fac2: Cumulative frequency (above) 
#> Fp: Percentage frequency 
#> Fac1p: Cumulative percentage frequency (below) 
#> Fac2p: Cumulative percentage frequency (above) 
#> 

# Example 2 (Pipe operator)
rnorm(36, 100, 4) |>
  new_leem(variable = "continuous") |> tabfreq()
#> 
#> Table of frequency 
#> Type of variable: continuous
#> 
#>               Classes Fi     PM   Fr Fac1 Fac2 Fp  Fac1p  Fac2p
#> 1    92.8 |---  96.18  4  94.49 0.11    4   36 11  11.11 100.00
#> 2   96.18 |---  99.56 11  97.87 0.31   15   32 31  41.67  88.89
#> 3  99.56 |---  102.94 12 101.25 0.33   27   21 33  75.00  58.33
#> 4 102.94 |---  106.32  5 104.63 0.14   32    9 14  88.89  25.00
#> 5  106.32 |---  109.7  3 108.01 0.08   35    4  8  97.22  11.11
#> 6  109.7 |---  113.08  1 111.39 0.03   36    1  3 100.00   2.78
#> 
#> ============================================== 
#> Classes: Grouping of classes 
#> Fi: Absolute frequency 
#> PM: Midpoint 
#> Fr: Relative frequency 
#> Fac1: Cumulative frequency (below) 
#> Fac2: Cumulative frequency (above) 
#> Fp: Percentage frequency 
#> Fac1p: Cumulative percentage frequency (below) 
#> Fac2p: Cumulative percentage frequency (above) 
#> 

# Example 3
x <- rbinom(36, 10, 0.6)
# Constructor (object of leem class)
x <- new_leem(x, variable = "discrete")
tab <- tabfreq(x)
# Details
tab$table
#>   Groups Fi   Fr Fac1 Fac2 Fp  Fac1p  Fac2p
#> 1      4  4 0.11    4   36 11  11.11 100.00
#> 2      5  9 0.25   13   32 25  36.11  88.89
#> 3      6  8 0.22   21   23 22  58.33  63.89
#> 4      7  7 0.19   28   15 19  77.78  41.67
#> 5      8  7 0.19   35    8 19  97.22  22.22
#> 6      9  1 0.03   36    1  3 100.00   2.78
tab$statistics
#> $ngroups
#> [1] 6
#> 
#> $minv
#> [1] 4
#> 
#> $maxv
#> [1] 9
#> 
#> $raw_data
#>  [1] 6 5 7 6 8 6 6 4 5 7 6 8 6 8 7 7 5 9 5 8 5 8 6 8 4 4 5 6 5 4 7 5 7 8 7 5
#> 

# Example 3 - ordered categories ("d","a", "b", "c")
w <- rep(letters[1:4], 1:4)
w |> new_leem(variable = "discrete") |> tabfreq(ordered = c("d","a", "b", "c"))
#> 
#> Table of frequency 
#> Type of variable: discrete
#> 
#>   Groups Fi  Fr Fac1 Fac2 Fp Fac1p Fac2p
#> 1      d  4 0.4    4   10 40    40   100
#> 2      a  1 0.1    5    6 10    50    60
#> 3      b  2 0.2    7    5 20    70    50
#> 4      c  3 0.3   10    3 30   100    30
#> ============================================== 
#> Groups: Discretized grouping 
#> Fi: Absolute frequency 
#> Fr: Relative frequency 
#> Fac1: Cumulative frequency (below) 
#> Fac2: Cumulative frequency (above) 
#> Fp: Percentage frequency 
#> Fac1p: Cumulative percentage frequency (below) 
#> Fac2p: Cumulative percentage frequency (above) 
#>