Frequency distribution table

Allows you to tabulate continuous and categorical data (quantitative or qualitative) in frequency distribution. Depending on the nature of the data, they can be grouped into class ranges or not.

Usage

# S3 method for class 'leem'
tabfreq(
  data,
  k = NULL,
  na.rm = FALSE,
  ordered = NULL,
  namereduction = TRUE,
  ...
)

Arguments

data

R object (data structure vector) of class leem. Use new_leem() function.

k

Number of classes. Default is NULL.

na.rm

a logical evaluating to TRUE or FALSE indicating whether NA values should be stripped before the computation proceeds.

ordered

Ordered vector of the same length and elements of data object. Default is NULL.

namereduction

Logical argument. If TRUE (default), the group names are reduzed the 10 characters. If FALSE, otherwise.

...

further arguments passed to or from other methods.

Value

The result of tabfreq() is a list. This list has two elements: table and statistics. The first is the data frequency table, and the second represents some useful statistics for methods of leem class.

Examples

# Example 1
library(leem)
x <- rbinom(36, 10, 0.6)
x <- new_leem(x, variable = "discrete")
tabfreq(x)
#> 
#> Table of frequency 
#> Type of variable: discrete
#> 
#>   Groups Fi   Fr Fac1 Fac2 Fp  Fac1p  Fac2p
#> 1      1  1 0.03    1   36  3   2.78 100.00
#> 2      2  1 0.03    2   35  3   5.56  97.22
#> 3      3  1 0.03    3   34  3   8.33  94.44
#> 4      4  4 0.11    7   33 11  19.44  91.67
#> 5      5  5 0.14   12   29 14  33.33  80.56
#> 6      6  7 0.19   19   24 19  52.78  66.67
#> 7      7  7 0.19   26   17 19  72.22  47.22
#> 8      8  8 0.22   34   10 22  94.44  27.78
#> 9      9  2 0.06   36    2  6 100.00   5.56
#> ============================================== 
#> Groups: Discretized grouping 
#> Fi: Absolute frequency 
#> Fr: Relative frequency 
#> Fac1: Cumulative frequency (below) 
#> Fac2: Cumulative frequency (above) 
#> Fp: Percentage frequency 
#> Fac1p: Cumulative percentage frequency (below) 
#> Fac2p: Cumulative percentage frequency (above) 
#> 

# Example 2 (Pipe operator)
rnorm(36, 100, 4) |>
  new_leem(variable = "continuous") |> tabfreq()
#> 
#> Table of frequency 
#> Type of variable: continuous
#> 
#>               Classes Fi     PM   Fr Fac1 Fac2 Fp  Fac1p  Fac2p
#> 1   89.54 |---  93.14  3  91.34 0.08    3   36  8   8.33 100.00
#> 2   93.14 |---  96.74  2  94.94 0.06    5   33  6  13.89  91.67
#> 3  96.74 |---  100.34 17  98.54 0.47   22   31 47  61.11  86.11
#> 4 100.34 |---  103.94  6 102.14 0.17   28   14 17  77.78  38.89
#> 5 103.94 |---  107.54  6 105.74 0.17   34    8 17  94.44  22.22
#> 6 107.54 |---  111.14  2 109.34 0.06   36    2  6 100.00   5.56
#> 
#> ============================================== 
#> Classes: Grouping of classes 
#> Fi: Absolute frequency 
#> PM: Midpoint 
#> Fr: Relative frequency 
#> Fac1: Cumulative frequency (below) 
#> Fac2: Cumulative frequency (above) 
#> Fp: Percentage frequency 
#> Fac1p: Cumulative percentage frequency (below) 
#> Fac2p: Cumulative percentage frequency (above) 
#> 

# Example 3
x <- rbinom(36, 10, 0.6)
# Constructor (object of leem class)
x <- new_leem(x, variable = "discrete")
tab <- tabfreq(x)
# Details
tab$table
#>   Groups Fi   Fr Fac1 Fac2 Fp  Fac1p  Fac2p
#> 1      1  2 0.06    2   36  6   5.56 100.00
#> 2      2  2 0.06    4   34  6  11.11  94.44
#> 3      3  2 0.06    6   32  6  16.67  88.89
#> 4      4  2 0.06    8   30  6  22.22  83.33
#> 5      5  7 0.19   15   28 19  41.67  77.78
#> 6      6 10 0.28   25   21 28  69.44  58.33
#> 7      7  7 0.19   32   11 19  88.89  30.56
#> 8      9  3 0.08   35    4  8  97.22  11.11
#> 9     10  1 0.03   36    1  3 100.00   2.78
tab$statistics
#> $ngroups
#> [1] 9
#> 
#> $minv
#> [1] 1
#> 
#> $maxv
#> [1] 10
#> 
#> $raw_data
#>  [1]  5  5  3  5  1  9  6  6  6  4  1  5  3  6  5  9  5  6  9  7  2  6  7  7  6
#> [26]  7 10  6  5  2  6  4  6  7  7  7
#> 

# Example 3 - ordered categories ("d","a", "b", "c")
w <- rep(letters[1:4], 1:4)
w |> new_leem(variable = "discrete") |> tabfreq(ordered = c("d","a", "b", "c"))
#> 
#> Table of frequency 
#> Type of variable: discrete
#> 
#>   Groups Fi  Fr Fac1 Fac2 Fp Fac1p Fac2p
#> 1      d  4 0.4    4   10 40    40   100
#> 2      a  1 0.1    5    6 10    50    60
#> 3      b  2 0.2    7    5 20    70    50
#> 4      c  3 0.3   10    3 30   100    30
#> ============================================== 
#> Groups: Discretized grouping 
#> Fi: Absolute frequency 
#> Fr: Relative frequency 
#> Fac1: Cumulative frequency (below) 
#> Fac2: Cumulative frequency (above) 
#> Fp: Percentage frequency 
#> Fac1p: Cumulative percentage frequency (below) 
#> Fac2p: Cumulative percentage frequency (above) 
#>